BOX-97: refactor(card-edit): decompose by layers (to entities)
